Add span to snarkjs assertions.
We generate witness-calculation code which contains assertions
corresponding to constraints. These assertions take the form of a
2-argument library function. Now we add a third, string, argument which
indicates the source span that generated the assertion (e.g.
bitify.circom:35). This is helpful because when a witness cannot be
calculated due to a failed constraint, the programmer is given a
line/file to look at.

For this to work, snarkjs must actually use that third argument to
produce a better error.

This change is, on its own, backward compatible with old snarkjs
versions, because extra JS arguments are permitted and ignored.